@@ -1,11 +1,9 @@
 ﻿using Application.Exceptions;
 using Application.Interfaces;
 using EmbPortal.Shared.Enums;
-using EmbPortal.Shared.Responses;
 using Infrastructure.Interfaces;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -26,7 +24,7 @@ namespace Application.CQRS.MBSheets.Command
             _currentUserService = currentUserService;
         }
 
-        public async Task<Unit> Handle(AcceptMBSheetCommand request, CancellationToken cancellationToken)
+        public async Task Handle(AcceptMBSheetCommand request, CancellationToken cancellationToken)
         {
             var mbSheet = await _context.MBSheets.Include(m => m.Items)
                 .ThenInclude(p => p.Measurements)
@@ -55,7 +53,6 @@ namespace Application.CQRS.MBSheets.Command
             mbSheet.MarkAsAccepted();
             await _context.SaveChangesAsync(cancellationToken);
 
-            return Unit.Value;
         }
     }
 }
