@@ -23,7 +23,7 @@ namespace Application.CQRS.MBSheets.Command
             _context = context;
         }
 
-        public async Task<Unit> Handle(EditMBSheetCommand request, CancellationToken cancellationToken)
+        public async Task Handle(EditMBSheetCommand request, CancellationToken cancellationToken)
         {
             MBSheet mbSheet = await _context.MBSheets
                 .Include(p => p.Items)
@@ -38,7 +38,6 @@ namespace Application.CQRS.MBSheets.Command
 
             await _context.SaveChangesAsync(cancellationToken);
 
-            return Unit.Value;
         }
     }
 }
