@@ -63,7 +63,7 @@
     public MBItemMeasurementRequest model { get; set; } = new();
 
     private List<WorkOrderItemStatusResponse> mbItemsStatuses = new();
-    private WorkOrderItemStatusResponse selectedItem;
+    
     private string title = "Add detail measurements";
 
     protected override void OnParametersSet()
