@@ -7,10 +7,10 @@
        ScrollX="450"
        HidePagination>
 
-    <RowTemplate>
+    <ColumnDefinitions>
         <Column Title="Item No" @bind-Field="@context.ItemNo" Width="20%" />
         <Column Title="Description" @bind-Field="@context.ItemDescription" Width="80%" />
-    </RowTemplate>
+    </ColumnDefinitions>
 
     <ExpandTemplate Context="rowData">
         <Table TItem="RaItemView"
