@@ -8,15 +8,16 @@
        ScrollX="450"
        HidePagination>
 
-    <RowTemplate>
+    <ColumnDefinitions>
         <Column Title="Item No" @bind-Field="@context.ItemNo" Width="20%" />
         <Column Title="Description" @bind-Field="@context.Description" Width="80%" />
-    </RowTemplate>
+    </ColumnDefinitions>
 
     <ExpandTemplate Context="rowData">
         <Table TItem="WorkOrderItemResponse"
                DataSource="@Items.Where(p => p.ItemNo == rowData.Data.ItemNo)"
                Bordered>
+
             <Column TData="int" Title="SubItem No" Width="15%">@context.SubItemNo</Column>
             <Column TData="int" Title="Service No" Width="15%">@context.ServiceNo</Column>
             <Column TData="string" Title="Description" Width="34%">@context.ShortServiceDesc</Column>
