@@ -1,5 +1,5 @@
-﻿@page "/"
-@using Client.Breadcrumbs;
+﻿@page "/" 
+
 @attribute [Authorize]
 <div class="text-center">    
     @if (currentUser != null){
