@@ -78,7 +78,7 @@ public class ExcelService : IExcelService
 
                 foreach (var value in result)
                 {
-                    ws.Cell(rowIndex, colIndex++).Value = value;
+                    ws.Cell(rowIndex, colIndex++).Value = (XLCellValue)value;
                 }
             }
             using (var stream = new MemoryStream())
